basic refactoring done, tab bar lacks style though
